backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# 测试依赖
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# 文档依赖
mkdocs==1.5.3
mkdocs-material==9.4.8